+++ src/main/java/bowling/model/Point.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.model;
 
 import java.util.Objects;
