+++ src/test/java/bowling/BowlingResultTest.java
@@ -1,5 +1,7 @@
 package bowling;
 
+import bowling.model.BowlingResult;
+import bowling.model.Point;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
