+++ src/test/java/bowling/FinalRoundTest.java
@@ -1,5 +1,7 @@
 package bowling;
 
+import bowling.model.BowlingResult;
+import bowling.model.FinalRound;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
