+++ src/test/java/bowling/PointTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.model.Point;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
