+++ src/test/java/bowling/UserTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.model.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
